Bump bitrate limits
This whole approach seems flawed to me.. the numbers in the docs for YouTube
aren't matching what I'm seeing at all, probably because they aren't maximums
but measurements from random content. It really seems like the max rate
we're setting for the encoder should function as a proxy for the expected
network and CPU resources available to the client more than anything else.